+++ layers/Engine/packages/root/src/AppThreadInterface.php
@@ -34,6 +34,21 @@ interface AppThreadInterface
     public function getName(): ?string;
 
     /**
+     * Store properties for identifying across different
+     * INTERNAL GraphQL servers, by storing the
+     * persisted query for each in the context.
+     *
+     * @return array<string,mixed>
+     */
+    public function getContext(): array;
+
+    /**
+     * Combination of the Name and Context
+     * to uniquely identify the AppThread
+     */
+    public function getUniqueID(): string;
+
+    /**
      * This function must be invoked at the very beginning,
      * to initialize the instance to run the application.
      *