+++ layers/GatoGraphQLForWP/plugins/gatographql/src/ConfigurationCache/InternalGraphQLServerContainerCacheConfigurationManager.php
@@ -4,14 +4,25 @@ declare(strict_types=1);
 
 namespace GatoGraphQL\GatoGraphQL\ConfigurationCache;
 
+use GatoGraphQL\GatoGraphQL\StateManagers\AppThreadHelper;
+
 class InternalGraphQLServerContainerCacheConfigurationManager extends ContainerCacheConfigurationManager
 {
+    /**
+     * @param array<string,mixed> $pluginAppGraphQLServerContext
+     */
+    public function __construct(
+        private array $pluginAppGraphQLServerContext
+    ) {
+    }
+
     /**
      * The internal server is always private, and has the
      * same configuration as the default admin endpoint.
      */
-    public function getNamespace(): string
+    protected function getNamespaceSuffix(): string
     {
-        return $this->makeNamespace($this->getNamespaceTimestampPrefix(), 'internal');
+        $graphQLServerContextID = AppThreadHelper::getGraphQLServerContextUniqueID($this->pluginAppGraphQLServerContext);
+        return 'internal_' . $graphQLServerContextID;
     }
 }