+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Container/InternalGraphQLServerContainerBuilderFactory.php
@@ -8,8 +8,21 @@ use PoP\Root\Container\ContainerBuilderFactory;
 
 class InternalGraphQLServerContainerBuilderFactory extends ContainerBuilderFactory
 {
+    use InternalGraphQLServerContainerBuilderFactoryTrait;
+
+    /**
+     * @param array<string,mixed> $pluginAppGraphQLServerContext
+     */
+    public function __construct(
+        private array $pluginAppGraphQLServerContext,
+    ) {
+    }
+
     public function getContainerClassName(): string
     {
-        return 'InternalGraphQLServer' . parent::getContainerClassName();
+        return $this->getInternalGraphQLServerContainerClassName(
+            $this->pluginAppGraphQLServerContext,
+            parent::getContainerClassName()
+        );
     }
 }