+++ layers/GatoGraphQLForWP/plugins/gatographql/src/PluginSkeleton/MainPluginInitializationConfigurationInterface.php
@@ -10,8 +10,11 @@ interface MainPluginInitializationConfigurationInterface extends PluginInitializ
 {
     /**
      * Provide the configuration to cache the container
+     *
+     * @param array<string,mixed> $pluginAppGraphQLServerContext
      */
     public function getContainerCacheConfiguration(
         string $pluginAppGraphQLServerName,
+        array $pluginAppGraphQLServerContext,
     ): ContainerCacheConfiguration;
 }