+++ layers/GatoGraphQLForWP/plugins/gatographql/src/StateManagers/AppThreadContext.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GatoGraphQL\GatoGraphQL\StateManagers;
+
+class AppThreadContext
+{
+    public final const SCHEMA_CONFIGURATION_ID = 'schemaConfigurationID';
+}