+++ layers/GraphQLByPoP/packages/graphql-server/src/Server/AbstractAttachedGraphQLServer.php
@@ -17,8 +17,9 @@ abstract class AbstractAttachedGraphQLServer extends AbstractGraphQLServer
     /**
      * Initialize the App with a new AppThread
      */
-    public function __construct()
-    {
+    public function __construct(
+        protected ?int $schemaConfigurationID
+    ) {
         /**
          * Steps:
          *